# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: trade-app-backend API
  version: 0.1.0
  description: >-
    课设精简版二手交易平台后端 API，覆盖认证、用户资料、分类、商品与最小订单闭环。
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      summary: Liveness health check
      responses:
        '200':
          description: ok
  /readyz:
    get:
      summary: Readiness health check
      responses:
        '200':
          description: ready
  /api/v1/auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEnvelope'
  /api/v1/auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEnvelope'
  /api/v1/users/me:
    get:
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelope'
    patch:
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeRequest'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelope'
  /api/v1/categories:
    get:
      summary: List categories
      responses:
        '200':
          description: Category list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoriesEnvelope'
  /api/v1/listings:
    get:
      summary: List published listings
      parameters:
        - in: query
          name: keyword
          schema:
            type: string
        - in: query
          name: category_id
          schema:
            type: integer
            format: int64
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Listing page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingsPageEnvelope'
    post:
      summary: Create listing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateListingRequest'
      responses:
        '200':
          description: Created listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingEnvelope'
  /api/v1/listings/{id}:
    get:
      summary: Get listing detail
      parameters:
        - $ref: '#/components/parameters/ListingID'
      responses:
        '200':
          description: Listing detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingEnvelope'
    patch:
      summary: Update listing
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ListingID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateListingRequest'
      responses:
        '200':
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingEnvelope'
  /api/v1/users/me/listings:
    get:
      summary: List current user's listings
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: My listings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingsPageEnvelope'
  /api/v1/orders:
    post:
      summary: Create order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '200':
          description: Created order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderEnvelope'
  /api/v1/orders/{id}/cancel:
    post:
      summary: Cancel order
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderID'
      responses:
        '200':
          description: Cancelled order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderEnvelope'
  /api/v1/orders/{id}/complete:
    post:
      summary: Complete order
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderID'
      responses:
        '200':
          description: Completed order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderEnvelope'
  /api/v1/users/me/orders:
    get:
      summary: List current user's orders
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: My orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdersPageEnvelope'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ListingID:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
    OrderID:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    SuccessMeta:
      type: object
      properties:
        code:
          type: string
          example: OK
        message:
          type: string
          example: success
    RegisterRequest:
      type: object
      required: [email, password, nickname]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        nickname:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    UpdateMeRequest:
      type: object
      properties:
        nickname:
          type: string
        bio:
          type: string
    CreateListingRequest:
      type: object
      required: [title, description, price_cents]
      properties:
        category_id:
          type: integer
          format: int64
          nullable: true
        title:
          type: string
        description:
          type: string
        price_cents:
          type: integer
          format: int64
        publish:
          type: boolean
    UpdateListingRequest:
      type: object
      required: [title, description, price_cents]
      properties:
        category_id:
          type: integer
          format: int64
          nullable: true
        title:
          type: string
        description:
          type: string
        price_cents:
          type: integer
          format: int64
        status:
          type: string
          enum: [draft, published, reserved, sold, archived]
    CreateOrderRequest:
      type: object
      required: [listing_id]
      properties:
        listing_id:
          type: integer
          format: int64
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        nickname:
          type: string
        bio:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AuthData:
      type: object
      properties:
        access_token:
          type: string
        user:
          $ref: '#/components/schemas/User'
    Listing:
      type: object
      properties:
        id:
          type: integer
          format: int64
        seller_id:
          type: integer
          format: int64
        category_id:
          type: integer
          format: int64
          nullable: true
        title:
          type: string
        description:
          type: string
        price_cents:
          type: integer
          format: int64
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
        listing_id:
          type: integer
          format: int64
        buyer_id:
          type: integer
          format: int64
        seller_id:
          type: integer
          format: int64
        status:
          type: string
          enum: [created, cancelled, completed]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AuthEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessMeta'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AuthData'
    UserEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessMeta'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/User'
    ListingEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessMeta'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Listing'
    OrderEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessMeta'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Order'
    CategoriesEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessMeta'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Category'
    ListingsPage:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/Listing'
    OrdersPage:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/Order'
    ListingsPageEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessMeta'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ListingsPage'
    OrdersPageEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessMeta'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OrdersPage'